scraper/requirements.txt
scrapy
mysql-connector-python
itemadapter
